remove simple-spellchecker
