+++ package.json
@@ -16,19 +16,10 @@
     "url": "git://github.com/mattermost/desktop.git"
   },
   "scripts": {
-    "postinstall": "electron-builder install-app-deps && npm run extract-dict",
-    "extract-dict": "node scripts/extract_dict.js src/node_modules/simple-spellchecker/dict",
-    "build": "npm-run-all build:*",
-    "build:main": "webpack-cli --bail --config webpack.config.main.js",
-    "build:renderer": "webpack-cli --bail --config webpack.config.renderer.js",
-    "start": "electron src --disable-dev-mode",
-    "restart": "npm run build && npm run start",
+    "build": "electron-webpack",
+    "start": "electron-webpack dev",
     "storybook": "start-storybook -p 9001 -c src/.storybook",
-    "clean": "rm -rf release/ node_modules/ src/node_modules/ && find src -name '*_bundle.js' | xargs rm",
-    "clean-install": "npm run clean && npm install",
-    "watch": "run-p watch:*",
-    "watch:main": "node scripts/watch_main_and_preload.js",
-    "watch:renderer": "webpack-dev-server --config webpack.config.renderer.js",
+    "clean": "rm -rf release/ node_modules/ && find src -name '*_bundle.js' | xargs rm",
     "test": "npm-run-all lint:js test:*",
     "test:app": "cross-env NODE_ENV=production npm run build && mocha -r @babel/register --reporter mocha-circleci-reporter --recursive test/specs",
     "package:all": "cross-env NODE_ENV=production npm-run-all check-build-config package:windows package:mac package:linux",
@@ -90,7 +81,7 @@
     "auto-launch": "^5.0.5",
     "bootstrap": "^3.3.7",
     "electron-context-menu": "^0.16.0",
-    "electron-devtools-installer": "^2.2.4",
+    "electron-devtools-installer": "^3.1.1",
     "electron-is-dev": "^1.0.1",
     "electron-log": "^4.1.3",
     "electron-updater": "4.0.6",
@@ -101,17 +92,20 @@
     "react-dom": "^16.6.3",
     "react-transition-group": "^2.5.0",
     "semver": "^5.5.0",
-    "simple-spellchecker": "^1.0.0",
     "underscore": "^1.9.1",
     "valid-url": "^1.0.9",
     "winreg": "^1.2.4",
     "yargs": "^15.3.1"
   },
   "electronWebpack": {
+    "staticSourceDirectory": "src/assets",
     "main": {
       "extraEntries": [
         "@/preload/mattermost.js"
       ]
+    },
+    "renderer": {
+      "webpackConfig": "webpack.config.renderer.js"
     }
   }
 }
