+++ src/main/preload/mattermost.js
@@ -221,34 +221,6 @@ function isElementVisible(elem) {
   return elem.offsetHeight !== 0;
 }
 
-function resetMisspelledState() {
-  ipcRenderer.once('spellchecker-is-ready', () => {
-    const element = document.activeElement;
-    if (element) {
-      element.blur();
-      element.focus();
-    }
-  });
-  ipcRenderer.send('reply-on-spellchecker-is-ready');
-}
-
-function setSpellChecker() {
-  const spellCheckerLocale = ipcRenderer.sendSync('get-spellchecker-locale');
-  webFrame.setSpellCheckProvider(spellCheckerLocale, {
-    spellCheck(words, callback) {
-      const misspeltWords = words.filter((text) => {
-        const res = ipcRenderer.sendSync('checkspell', text);
-        const isCorrect = (res === null) ? true : res;
-        return !isCorrect;
-      });
-      callback(misspeltWords);
-    },
-  });
-  resetMisspelledState();
-}
-setSpellChecker();
-ipcRenderer.on('set-spellchecker', setSpellChecker);
-
 // push user activity updates to the webapp
 ipcRenderer.on('user-activity-update', (event, {userIsActive, isSystemEvent}) => {
   if (window.location.origin !== 'null') {
