+++ src/renderer/components/MainPage.jsx
@@ -1,6 +1,6 @@
-// Copyright (c) 2015-2016 Yuya Ochiai
 // Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
+// Copyright (c) 2015-2016 Yuya Ochiai
 
 // This files uses setState().
 /* eslint-disable react/no-set-state */
@@ -16,8 +16,8 @@ import DotsVerticalIcon from 'mdi-react/DotsVerticalIcon';
 
 import {ipcRenderer, remote, shell} from 'electron';
 
-import Utils from '../../utils/util';
-import contextmenu from '../js/contextMenu';
+// eslint-disable-next-line import/no-unresolved
+import Utils from 'common/utils/util';
 
 import restoreButton from '../../assets/titlebar/chrome-restore.svg';
 import maximizeButton from '../../assets/titlebar/chrome-maximize.svg';
@@ -63,14 +63,6 @@ export default class MainPage extends React.Component {
       focusFinder: false,
       finderVisible: false,
     };
-    contextmenu.setup({
-      useSpellChecker: this.props.useSpellChecker,
-      onSelectSpellCheckerLocale: (locale) => {
-        if (this.props.onSelectSpellCheckerLocale) {
-          this.props.onSelectSpellCheckerLocale(locale);
-        }
-      },
-    });
   }
 
   parseDeeplinkURL(deeplink, teams = this.props.teams) {
