+++ src/renderer/components/SettingsPage.jsx
@@ -1,6 +1,6 @@
-// Copyright (c) 2015-2016 Yuya Ochiai
 // Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
+// Copyright (c) 2015-2016 Yuya Ochiai
 
 // This file uses setState().
 /* eslint-disable react/no-set-state */
