+++ webpack.config.renderer.js
@@ -4,38 +4,14 @@
 
 // This file uses CommonJS.
 /* eslint-disable import/no-commonjs */
-'use strict';
 
-const path = require('path');
-
-const merge = require('webpack-merge');
-
-const base = require('./webpack.config.base');
-
-module.exports = merge(base, {
-  entry: {
-    index: './src/browser/index.jsx',
-    settings: './src/browser/settings.jsx',
-    updater: './src/browser/updater.jsx',
-    'webview/mattermost': './src/browser/webview/mattermost.js',
-  },
-  output: {
-    path: path.join(__dirname, 'src/browser'),
-    publicPath: 'browser',
-    filename: '[name]_bundle.js',
-  },
+module.exports = {
   module: {
     rules: [{
       test: /\.(js|jsx)?$/,
       use: {
         loader: 'babel-loader',
       },
-    }, {
-      test: /\.css$/,
-      use: [
-        {loader: 'style-loader'},
-        {loader: 'css-loader'},
-      ],
     }, {
       test: /\.mp3$/,
       use: {
@@ -58,21 +34,9 @@ module.exports = merge(base, {
       loader: 'file-loader',
       options: {
         name: '[name].[ext]',
-        outputPath: '/../assets/fonts',
-        publicPath: './assets/fonts',
       },
     }],
   },
-  node: {
-    __filename: true,
-    __dirname: true,
-  },
-  target: 'electron-renderer',
-  devServer: {
-    contentBase: path.join(__dirname, 'src'),
-    inline: true,
-    publicPath: '/browser/',
-  },
-});
+};
 
 /* eslint-enable import/no-commonjs */
